match the parser's extern declarations to their definitions
parser.c declared error_count as int while error.c defines it as short, and
linenum as long while lex.yy.c defines it as int. On 32-bit MSVC int and long
are the same size, and the bytes after error_count happened to be zero, so it
never showed. On the static arm64 build warning_count sat right after
error_count, so the parser's `if (error_count == 0)` read any warning as an
error: a program with a warning compiled to nothing, with exit status 0 and
no message, while the same source and binary on x86-64 produced M.class.

Found by the new Selftest, which deliberately emits W464, failing the
x86-vs-arm64 diff with "Only in ...: M.class". The shipped android arm64
binary was not affected by this one — a different layout — but the fix
removes the luck from both.

+++ parser/parser.c
@@ -48,10 +48,10 @@ block *root_block;
 int next_record_ID;
 
 /* Scanner variables and functions */
-extern long int linenum;
+extern int linenum;          // int in lex.yy.c: a long here read 8 bytes of a 4-byte variable on lp64
 extern int new_linenum;
 extern long int integer_constant;
-extern int error_count;
+extern short int error_count; // short in error.c: an int here also read warning_count next to it
 extern float real_constant;
 extern char boolean_constant;
 extern char char_constant;
